pyproject.toml
[tool.poetry]
name = "eudi-wallet"
version = "0.0.1"
description = "EUDI Wallet SDK"
authors = ["George J Padayatti <george.padayatti@igrant.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
jwcrypto = "^1.3.1"
py-multibase = "^1.0.3"
coincurve = "^17.0.0"
aiohttp = "^3.8.1"
sslcrypto = "^5.3"
rich = "^12.5.1"
jsonschema = "^4.14.0"
base58 = "1.0.3"
eth-keys = "^0.4.0"
PyJWT = "^2.4.0"
pysha3 = "^1.0.2"
multiformats = "^0.2.1"
cryptography = "^41.0.1"
pyngrok = "^6.0.0"
click = "^8.1.3"
dataclasses-json = "^0.5.9"
web3 = "^6.5.0"
sqlalchemy = "^2.0.19"
pytz = "^2023.3"
confluent-kafka = "2.1.1"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
notebook = "^6.4.12"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disable_error_code = ["return"]